third_party/WebKit/LayoutTests/http/tests/htmlimports/redirect-expected.txt

CONSOLE ERROR: Imported resource from origin 'http://localhost:8000' has been blocked from loading by Cross-Origin Resource Sharing policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Imported resource from origin 'http://localhost:8000' has been blocked from loading by Cross-Origin Resource Sharing policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
This is a testharness.js-based test.
PASS Accessing to a same origin import 
PASS Accessing to CORS-enabled origin import 
PASS Blocking access to a cross origin import 
PASS Accessing to a same origin import including same domain redirect 
PASS Accessing to a CORS origin import including same domain redirect 
PASS Blocking access to a cross origin iamport including same domain redirect 
Harness: the test ran to completion.
